% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pooled_polr_test.R
\name{pooled_polr_test}
\alias{pooled_polr_test}
\title{Test an orderd logistic or probit regression for the presence of time-varying effects}
\usage{
pooled_polr_test(formula, Panel.data, N = 100, method = c("logistic",
  "probit", "loglog", "cloglog", "cauchit"))
}
\arguments{
\item{formula}{An object of class "formula", which provides a symbolic
description of the model to be fitted. Ensure that names included here are
contained with the data frame entries of the Panel.data object.}

\item{Panel.data}{A list of data frames containing the predictors and
response variables listed in the model formula. List entries are assumed
to be ordered according to time.}

\item{N}{The number of bootstrap samples to run for this test. Default is 100.}

\item{method}{Logistic, probit, log-log, or cauchit}
}
\description{
\code{pooled_polr_test} is used to test the temporal effect homogeneity
of a ordered logistic or probit regression with panel data.
}
